removed unused imports, changed file and directory structure and changed imports accordingly, added routing to App.jsx. added .gitignore
