--- main.jsx
@@ -1,21 +1,0 @@
-
-
-
-import { StrictMode } from 'react';
-import { createRoot } from 'react-dom/client';
-import './index.css';
-import App from './App.jsx';
-import { BrowserRouter } from 'react-router-dom';
-
-const rootElement = document.getElementById('root');
-const root = createRoot(rootElement);
-
-root.render(
-  <StrictMode>
-    <BrowserRouter>
-      
-        <App />
-      
-    </BrowserRouter>
-  </StrictMode>
-);
